findata-service/requirements.txt
fastapi>=0.115.0
uvicorn>=0.35.0
pydantic>=2.0.0
pandas>=2.0.0
requests>=2.32.0
lixinger-openapi>=0.1.0
cachetools>=5.5.0
python-dotenv>=1.0.0
